+++ apps/connect/src/context/processor-manager.tsx
@@ -4,16 +4,16 @@ import {
     AnalyticsProvider,
     useAnalyticsStoreAsync,
 } from '@powerhousedao/reactor-browser/analytics/context';
-import { useOperationalStore } from '@powerhousedao/reactor-browser/operational';
 import {
     live,
     useSetPGliteDB,
     type PGlite,
     type PGliteWithLive,
 } from '@powerhousedao/reactor-browser/pglite';
+import { useRelationalDb } from '@powerhousedao/reactor-browser/relational';
 import { childLogger } from 'document-drive';
 import type { ProcessorManager } from 'document-drive/processors/processor-manager';
-import { type IOperationalStore } from 'document-drive/processors/types';
+import { type IRelationalDb } from 'document-drive/processors/types';
 import { useEffect, useRef, type PropsWithChildren } from 'react';
 import {
     useExternalProcessors,
@@ -63,13 +63,13 @@ function createPgLiteFactoryWorker(databaseName: string) {
 async function registerExternalProcessors(
     manager: ProcessorManager,
     analyticsStore: IAnalyticsStore,
-    operationalStore: IOperationalStore,
+    relationalDb: IRelationalDb,
     processorName: string,
     processorFactory: Processors,
 ) {
     return await manager.registerFactory(
         processorName,
-        processorFactory({ analyticsStore, operationalStore }),
+        processorFactory({ analyticsStore, relationalDb }),
     );
 }
 
@@ -142,7 +142,7 @@ export function DriveAnalyticsProcessor() {
 export function ExternalProcessors() {
     const externalProcessors = useExternalProcessors();
     const store = useAnalyticsStoreAsync();
-    const operationalStore = useOperationalStore();
+    const relationalDb = useRelationalDb();
     const manager = useUnwrappedProcessorManager();
     const hasRegistered = useRef(false);
 
@@ -152,7 +152,7 @@ export function ExternalProcessors() {
             !manager ||
             hasRegistered.current ||
             externalProcessors.length === 0 ||
-            !operationalStore.db
+            !relationalDb.db
         ) {
             return;
         }
@@ -164,14 +164,14 @@ export function ExternalProcessors() {
             registerExternalProcessors(
                 manager,
                 store.data,
-                operationalStore.db,
+                relationalDb.db,
                 `${packageName}-${index}`,
                 processors,
             ).catch(logger.error);
 
             index++;
         }
-    }, [store.data, manager, operationalStore]);
+    }, [store.data, manager, relationalDb]);
 
     return null;
 }