+++ packages/codegen/src/codegen/.hygen/templates/powerhouse/generate-document-model-subgraph/resolvers.esm.t
@@ -7,6 +7,7 @@ force: true
 /* eslint-disable @typescript-eslint/no-unsafe-member-access */
 import { type Subgraph } from "@powerhousedao/reactor-api";
 import { addFile } from "document-drive";
+import { DocumentNotFoundError } from "document-drive/server/error";
 import { actions } from "../../document-models/<%- h.changeCase.param(documentType) %>/index.js";
 import { generateId } from "document-model";
 
@@ -23,32 +24,46 @@ export const getResolvers = (subgraph: Subgraph): Record<string, any> => {
             const driveId: string = args.driveId || DEFAULT_DRIVE_ID;
             const docId: string = args.docId || "";
             const doc = await reactor.getDocument(driveId, docId);
-             return {
+            if (!doc) {
+              throw new DocumentNotFoundError(docId);
+            }
+
+            return {
               driveId: driveId,
               ...doc,
+              ...doc.header,
               state: doc.state.global,
               stateJSON: doc.state.global,
-              revision: doc.revision.global,
+              revision: doc.header.revision["global"] ?? 0,
             };
           },
           getDocuments: async (args: any) => {
             const driveId: string = args.driveId || DEFAULT_DRIVE_ID;
             const docsIds = await reactor.getDocuments(driveId);
+            if (!docsIds) {
+              throw new DocumentNotFoundError(driveId);
+            }
+
             const docs = await Promise.all(
               docsIds.map(async (docId) => {
                 const doc = await reactor.getDocument(driveId, docId);
+                if (!doc) {
+                  return null;
+                }
+
                 return {
                   driveId: driveId,
                   ...doc,
+                  ...doc.header,
                   state: doc.state.global,
                   stateJSON: doc.state.global,
-                  revision: doc.revision.global,
+                  revision: doc.header.revision["global"] ?? 0,
                 };
               }),
             );
 
             return docs.filter(
-              (doc) => doc.documentType === "<%- documentTypeId %>",
+              (doc) => doc && doc.header.documentType === "<%- documentTypeId %>",
             );
           },
         };
@@ -88,6 +103,9 @@ export const getResolvers = (subgraph: Subgraph): Record<string, any> => {
             const driveId: string = args.driveId || DEFAULT_DRIVE_ID;
             const docId: string = args.docId || "";
             const doc = await reactor.getDocument(driveId, docId);
+            if (!doc) {
+              throw new DocumentNotFoundError(docId);
+            }
 
             await reactor.addAction(
                 driveId,
@@ -95,7 +113,7 @@ export const getResolvers = (subgraph: Subgraph): Record<string, any> => {
                 actions.<%- h.changeCase.camel(op.name) %>({...args.input})
             );
             
-            return doc.revision.global + 1;
+            return (doc.header.revision["global"] ?? 0) + 1;
         },
 
 <%_ })}); %>