+++ packages/codegen/src/codegen/.hygen/templates/powerhouse/generate-processor-relationalDb/migrations.esm.t
@@ -2,9 +2,9 @@
 to: "<%= rootDir %>/<%= h.changeCase.param(name) %>/migrations.ts"
 force: true
 ---
-import { type IOperationalStore } from "document-drive/processors/types"
+import { type IRelationalDb } from "document-drive/processors/types"
 
-export async function up(db: IOperationalStore): Promise<void> {
+export async function up(db: IRelationalDb<any>): Promise<void> {
   // Create table 
   await db.schema
     .createTable("todo")
@@ -17,7 +17,7 @@ export async function up(db: IOperationalStore): Promise<void> {
     .execute();
 }
 
-export async function down(db: IOperationalStore): Promise<void> {
+export async function down(db: IRelationalDb<any>): Promise<void> {
   // drop table
   await db.schema.dropTable("todo").execute();
 }