+++ packages/common/state/atoms.ts
@@ -10,7 +10,7 @@ import {
   type EditorModule,
   type PHDocument,
 } from "document-model";
-import { atom, type WritableAtom } from "jotai";
+import { atom } from "jotai";
 import { atomWithRefresh, loadable, unwrap } from "jotai/utils";
 import { isFolderNodeKind } from "./nodes.js";
 import type { PHPackage } from "./ph-packages.js";
@@ -70,15 +70,10 @@ unwrappedReactorAtom.debugLabel = "unwrappedReactorAtom";
  *
  * Suspends until the reactor is set.
  *
- * Does not provide a direct setter, isnstead it uses `atomWithRefresh` which will refetch the drives from the reactor when called.
+ * Does not provide a direct setter, instead it uses `atomWithRefresh` which will refetch the drives from the reactor when called.
  * See https://jotai.org/docs/utilities/resettable#atomwithrefresh for more details.
  */
-export const drivesAtom: WritableAtom<
-  Promise<DocumentDriveDocument[]>,
-  // eslint-disable-next-line @typescript-eslint/no-invalid-void-type
-  [void],
-  void
-> = atomWithRefresh(async (get) => {
+export const drivesAtom = atomWithRefresh(async (get) => {
   const loadableReactor = get(loadableReactorAtom);
 
   // Suspends until the reactor is set.
@@ -91,9 +86,12 @@ export const drivesAtom: WritableAtom<
   if (!reactor) return [];
 
   const driveIds = await reactor.getDrives();
-  const drives = (
-    await Promise.all(driveIds.map((id) => reactor.getDrive(id)))
-  ).filter((drive): drive is DocumentDriveDocument => Boolean(drive));
+  const drives = new Array<DocumentDriveDocument>();
+
+  for (const driveId of driveIds) {
+    const drive = await reactor.getDrive(driveId);
+    drives.push(drive);
+  }
 
   return drives;
 });