+++ packages/reactor-api/src/graphql/base/index.ts
@@ -1,7 +1,7 @@
 import { type GraphQLManager } from "#graphql/graphql-manager.js";
 import { type ISubgraph, type SubgraphArgs } from "#graphql/types.js";
-import { type Db } from "#types.js";
 import { type IDocumentDriveServer } from "document-drive";
+import { type IRelationalDb } from "document-drive/processors/types";
 import { type DocumentNode } from "graphql";
 import { gql } from "graphql-tag";
 
@@ -20,13 +20,15 @@ export class Subgraph implements ISubgraph {
   `;
   reactor: IDocumentDriveServer;
   graphqlManager: GraphQLManager;
-  operationalStore: Db;
+  relationalDb: IRelationalDb;
+
   constructor(args: SubgraphArgs) {
     this.reactor = args.reactor;
     this.graphqlManager = args.graphqlManager;
-    this.operationalStore = args.operationalStore;
+    this.relationalDb = args.relationalDb;
     this.path = args.path ?? "";
   }
+
   async onSetup() {
     // noop
   }